Renamed old api routes file
